Crate arctk

Library core.

Modules

data
Data reduction.

err
Error handling.

file
File IO.

geom
Three-dimensional Euclidean geometry.

img
Image tools module.

math
Mathematics.

meta
Meta-programming.

ord
Ordering module.

phys
Physics module.

sim
Simulation specific structures.

tools
Useful tools.

util
Utilities.

Macros

access
Succinctly create a reference method for the given variable.

args
Import command line arguments as a requested type.

clone
Succinctly create a clone method for the given variable.

fmt_report
Report an object and either its associated name, or a human readable string if supplied.

fmt_reports
Report an iterable object and either its associated name, or a human readable string if supplied.

map
Concisely construct a hashmap.

pause
Pause the thread for the given number of milliseconds.

report
Report an object and either its associated name, or a human readable string if supplied.

reports
Report an iterable object and either its associated name, or a human readable string if supplied.