Trait arctk::fs::file::File

pub trait File where
    Self: Sized, {
    fn load(path: &Path) -> Result<Self, Error>;

    fn new_from_file(path: &Path) -> Result<Self, Error> { ... }
}

Types implementing this trait can be loaded from a file.

Required methods

fn load(path: &Path) -> Result<Self, Error>

Deserialize the type from a given file.

Errors

if the target file can not be found, or the read string can not be serialised into an instance of the required type.

Loading content...
Provided methods

fn new_from_file(path: &Path) -> Result<Self, Error>

Load an instance of this type from a given path.

Errors

if the target file can not be found, or the read string can not be serialised into an instance of the required type.

Loading content...