Trait arctk::ord::name::Name

pub trait Name {
    type Inst;
    fn reg(self, reg: &Register) -> Result<Self::Inst, Error>;
}

Types implementing this trait can be built at runtime from an input structure with names rather than indices.

Associated Types

type Inst

Type to be constructed.

Loading content...
Required methods

fn reg(self, reg: &Register) -> Result<Self::Inst, Error>

Build the instance type by registering the names as indices.

Errors

if a component could not be named successfully.

Loading content...