Trait arctk::file::save::Save

pub trait Save {
    fn save(&self, path: &Path) -> Result<(), Error>;
}

Types implementing this trait can be saved to file.

Required methods

fn save(&self, path: &Path) -> Result<(), Error>

Serialise the type to a given file

Errors

if the instance can not be serialised or if the file can't be written to.

Loading content...